Update 3.6 Part1 (I fricked up...)

+++ Projectiles/NyivesBigProj.cs
@@ -36,24 +36,27 @@ namespace VampKnives.Projectiles
         }
 
         public override void SafeAI()
-        {
-            projectile.rotation = projectile.velocity.ToRotation() + MathHelper.ToRadians(SpriteRotation); // projectile faces sprite right
+        {            
             Lighting.AddLight(projectile.Center, (Main.DiscoR / 255f), (Main.DiscoG / 255f), (Main.DiscoB / 255f));
-            if (HitTile)
+            if (!ZenithActive)
             {
-                projectile.velocity = projectile.velocity * 0.55f;
-                if (HasDoneEffect == false)
+                projectile.rotation = projectile.velocity.ToRotation() + MathHelper.ToRadians(SpriteRotation); // projectile faces sprite right
+                if (HitTile)
                 {
-                    HasDoneEffect = true;
-                    for (int i = 0; i < 10; i++)
+                    projectile.velocity = projectile.velocity * 0.55f;
+                    if (HasDoneEffect == false)
                     {
-                        int DustID2 = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y), projectile.width - 3, projectile.height - 3, 1, projectile.velocity.X * 0.2f, projectile.velocity.Y * -0.2f, 10, Color.Gray, 1f);
+                        HasDoneEffect = true;
+                        for (int i = 0; i < 10; i++)
+                        {
+                            int DustID2 = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y), projectile.width - 3, projectile.height - 3, 1, projectile.velocity.X * 0.2f, projectile.velocity.Y * -0.2f, 10, Color.Gray, 1f);
+                        }
                     }
                 }
-            }
-            if(HitCount > 10)
-            {
-                projectile.Kill();
+                if (HitCount > 10)
+                {
+                    projectile.Kill();
+                }
             }
         }
         public override void ModifyHitNPC(NPC target, ref int damage, ref float knockback, ref bool crit, ref int hitDirection)
@@ -61,18 +64,18 @@ namespace VampKnives.Projectiles
             projectile.penetrate++;
             HitCount++;
         }
-        public override void OnHitNPC(NPC n, int damage, float knockback, bool crit)
-        {
-            Player owner = Main.player[projectile.owner];
-            if (HitCount == 1)
-            {
-                if (Main.rand.Next(0, HealProjChanceScale) <= HealProjChance)
-                {
-                    Projectile.NewProjectile(projectile.position.X, projectile.position.Y, 0, 0, mod.ProjectileType("HealProj"), (int)(projectile.damage * 0.75), 0, owner.whoAmI);
-                }
-                Hoods(n);
-            }
-        }
+        //public override void OnHitNPC(NPC n, int damage, float knockback, bool crit)
+        //{
+        //    Player owner = Main.player[projectile.owner];
+        //    if (HitCount == 1)
+        //    {
+        //        if (Main.rand.Next(0, HealProjChanceScale) <= HealProjChance)
+        //        {
+        //            Projectile.NewProjectile(projectile.position.X, projectile.position.Y, 0, 0, mod.ProjectileType("HealProj"), (int)(projectile.damage * 0.75), 0, owner.whoAmI);
+        //        }
+        //        Hoods(n);
+        //    }
+        //}
         public override bool SafeOnTileCollide(Vector2 oldVelocity)
         {
                 HitTile = true;
